requirements.txt
asgiref==3.7.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==5.0
django-allauth==0.57.0
django-colorfield==0.11.0
django-contrib-postgres==0.0.1
django-crispy-forms==2.1
django-summernote==0.8.20.0
fontawesomefree==6.5.1
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.6.0
